EC-322 - all tests now passing after the refactor with the ConsultationState
