+++ Comments.Test/IntegrationTests/API/Comments/CommentTests.Get_Comments_Feed_ReturnsEmptyFeed.approved.txt
@@ -8,16 +8,16 @@
     "isAuthorised":true,
     "signInURL":"/signin?returnURL=/",
     "consultationState":{
-        "startDate":"2018-04-01T00:00:00",
-        "endDate":"2018-05-31T17:00:00",
-        "hasQuestions":false,
-        "hasUserSuppliedAnswers":false,
-        "hasUserSuppliedComments":false,
-        "userHasSubmitted":false,
-        "consultationIsOpen":false,
+        "startDate":"0001-01-01T00:00:00",
+        "endDate":"9999-12-31T23:59:59.9999999",
+        "hasQuestions":true,
+        "hasUserSuppliedAnswers":true,
+        "hasUserSuppliedComments":true,
+        "userHasSubmitted":true,
+        "consultationIsOpen":true,
         "consultationHasNotStartedYet":false,
-        "consultationHasEnded":true,
+        "consultationHasEnded":false,
         "supportsSubmission":false,
-        "supportsDownload":false
+        "supportsDownload":true
     }
 }
