+++ Comments.Test/IntegrationTests/API/Comments/CommentTests.Insert_Multiple_Comments_And_Read_Them.approved.txt
@@ -67,15 +67,15 @@
     "isAuthorised":true,
     "signInURL":"/signin?returnURL=/",
     "consultationState":{
-        "startDate":"2018-04-01T00:00:00",
-        "endDate":"2018-05-31T17:00:00",
-        "hasQuestions":false,
-        "hasUserSuppliedAnswers":false,
+        "startDate":"0001-01-01T00:00:00",
+        "endDate":"9999-12-31T23:59:59.9999999",
+        "hasQuestions":true,
+        "hasUserSuppliedAnswers":true,
         "hasUserSuppliedComments":true,
-        "userHasSubmitted":false,
-        "consultationIsOpen":false,
+        "userHasSubmitted":true,
+        "consultationIsOpen":true,
         "consultationHasNotStartedYet":false,
-        "consultationHasEnded":true,
+        "consultationHasEnded":false,
         "supportsSubmission":false,
         "supportsDownload":true
     }
