+++ Comments.Test/IntegrationTests/API/Comments/CommentTests.cs
@@ -149,7 +149,9 @@ namespace Comments.Test.IntegrationTests.API.Comments
 
             var userService = FakeUserService.Get(isAuthenticated: true, displayName: "Benjamin Button", userId: userId);
             var authenticateService = new FakeAuthenticateService(authenticated: true);
-            var commentService = new CommentService(new ConsultationsContext(_options, userService), userService, authenticateService, _consultationService);
+	        var context = new ConsultationsContext(_options, userService);
+			var submitService = new SubmitService(context, userService, _consultationService);
+            var commentService = new CommentService(context, userService, authenticateService, submitService);
 
             //Act
             var response = await _client.DeleteAsync($"consultations/api/Comment/{commentId}");
@@ -176,7 +178,8 @@ namespace Comments.Test.IntegrationTests.API.Comments
 
             var userService = FakeUserService.Get(isAuthenticated: true, displayName: "Benjamin Button", userId: userId);
             var authenticateService = new FakeAuthenticateService(authenticated: true);
-            var commentService = new CommentService(_context, userService, authenticateService, _consultationService);
+	        var submitService = new SubmitService(_context, userService, _consultationService);
+			var commentService = new CommentService(_context, userService, authenticateService, submitService);
 
             var viewModel = commentService.GetComment(commentId);
