+++ Comments.Test/IntegrationTests/API/Submit/SubmitTests.cs
@@ -15,7 +15,7 @@ namespace Comments.Test.IntegrationTests.API.Submit
 {
     public class SubmitTests: TestBase
     {
-	    public SubmitTests() : base(useFakeConsultationService:true) {}
+	    public SubmitTests() : base(useRealSubmitService:true) {}
 
 	    [Fact]
 	    public async Task Submit_Comment_And_Answers()
@@ -36,7 +36,8 @@ namespace Comments.Test.IntegrationTests.API.Submit
 
 			var userService = FakeUserService.Get(isAuthenticated: true, displayName: "Benjamin Button", userId: userId);
 		    var authenticateService = new FakeAuthenticateService(authenticated: true);
-			var commentService = new CommentService(_context, userService, authenticateService, _consultationService);
+			var submitService = new SubmitService(_context, userService, _consultationService);
+			var commentService = new CommentService(_context, userService, authenticateService, submitService);
 
 		    var commentsAndQuestions = commentService.GetCommentsAndQuestions(sourceURI, true);
 		    var viewModel = new CommentsAndAnswers(commentsAndQuestions.Comments, commentsAndQuestions.Questions.First().Answers);
