+++ Comments.Test/UnitTests/CommentService.Tests.cs
@@ -31,7 +31,9 @@ namespace Comments.Test.UnitTests
 
             var userService = FakeUserService.Get(isAuthenticated: true, displayName: "Benjamin Button", userId: userId);
             var authenticateService = new FakeAuthenticateService(authenticated: true);
-            var commentService = new CommentService(new ConsultationsContext(_options, userService), userService, authenticateService, _consultationService);
+	        var context = new ConsultationsContext(_options, userService);
+	        var submitService = new SubmitService(context, userService, _consultationService);
+			var commentService = new CommentService(context, userService, authenticateService, submitService);
 
             // Act
             var viewModel = commentService.GetComment(commentId);
@@ -50,7 +52,9 @@ namespace Comments.Test.UnitTests
             var userService = FakeUserService.Get(isAuthenticated: true, displayName: "Benjamin Button", userId: userId);
             var authenticateService = new FakeAuthenticateService(authenticated: true);
 
-            var commentService = new CommentService(new ConsultationsContext(_options, userService), userService, authenticateService, _consultationService);
+	        var context = new ConsultationsContext(_options, userService);
+			var submitService = new SubmitService(context, userService, _consultationService);
+			var commentService = new CommentService(context, userService, authenticateService, submitService);
 
             // Act
             var viewModel = commentService.GetComment(1);
@@ -76,7 +80,9 @@ namespace Comments.Test.UnitTests
 
             var userService = FakeUserService.Get(isAuthenticated: true, displayName: "Benjamin Button", userId: userId);
             var authenticateService = new FakeAuthenticateService(authenticated: true);
-            var commentService = new CommentService(new ConsultationsContext(_options, userService), userService, authenticateService, _consultationService);
+	        var context = new ConsultationsContext(_options, userService);
+			var submitService = new SubmitService(context, userService, _consultationService);
+			var commentService = new CommentService(context, userService, authenticateService, submitService);
 
             var viewModel = commentService.GetComment(commentId);
             var updatedCommentText = Guid.NewGuid().ToString();
@@ -105,8 +111,10 @@ namespace Comments.Test.UnitTests
 
             var locationId = AddLocation(sourceURI);
             var commentId = AddComment(locationId, commentText, isDeleted: false, createdByUserId: userId);
-            
-            var commentService = new CommentService(new ConsultationsContext(_options, userService), userService, authenticateService, _consultationService);
+
+	        var context = new ConsultationsContext(_options, userService);
+	        var submitService = new SubmitService(context, userService, _consultationService);
+			var commentService = new CommentService(context, userService, authenticateService, submitService);
 
             //Act
             commentService.DeleteComment(commentId);
@@ -134,7 +142,9 @@ namespace Comments.Test.UnitTests
             
             var userService = FakeUserService.Get(isAuthenticated: true, displayName: "Benjamin Button", userId: userId);
             var authenticateService = new FakeAuthenticateService(authenticated: true);
-            var commentService = new CommentService(new ConsultationsContext(_options, userService), userService, authenticateService, _consultationService);
+	        var context = new ConsultationsContext(_options, userService);
+			var submitService = new SubmitService(context, userService, _consultationService);
+			var commentService = new CommentService(context, userService, authenticateService, submitService);
 
             //Act
             var result = commentService.CreateComment(viewModel);
@@ -149,7 +159,9 @@ namespace Comments.Test.UnitTests
             // Arrange
             ResetDatabase();
             var authenticateService = new FakeAuthenticateService(authenticated: false);
-            var commentService = new CommentService(new ConsultationsContext(_options, _fakeUserService), FakeUserService.Get(isAuthenticated: false), authenticateService, _consultationService);
+	        var context = new ConsultationsContext(_options, _fakeUserService);
+	        var submitService = new SubmitService(context, _fakeUserService, _consultationService);
+			var commentService = new CommentService(new ConsultationsContext(_options, _fakeUserService), FakeUserService.Get(isAuthenticated: false), authenticateService, submitService);
 
             // Act
             var viewModel = commentService.GetCommentsAndQuestions("consultations://./consultation/1/document/1/chapter/introduction");
@@ -171,8 +183,9 @@ namespace Comments.Test.UnitTests
             var sourceURI = "consultations://./consultation/1/document/1/chapter/introduction";
             var userService = FakeUserService.Get(isAuthenticated: true, displayName: "Benjamin Button", userId: userId);
             var authenticateService = new FakeAuthenticateService(authenticated: true);
-
-            var commentService = new CommentService(new ConsultationsContext(_options, userService), userService, authenticateService, _consultationService);
+	        var context = new ConsultationsContext(_options, userService);
+	        var submitService = new SubmitService(context, userService, _consultationService);
+			var commentService = new CommentService(new ConsultationsContext(_options, userService), userService, authenticateService, submitService);
             var locationId = AddLocation(sourceURI);
 
             var expectedCommentId = AddComment(locationId, "current user's comment", isDeleted: false, createdByUserId: userId);
@@ -204,8 +217,9 @@ namespace Comments.Test.UnitTests
 
             AddCommentsAndQuestionsAndAnswers(URI, "My Comment", someText, someText, createdByUserId, (int)StatusName.Draft, _context);
             AddCommentsAndQuestionsAndAnswers(URI, "Another Users Comment", someText, someText, anotherUserId, (int)StatusName.Draft, _context);
-
-            var commentService = new CommentService(new ConsultationsContext(_options, userService), _fakeUserService, authenticateService, _consultationService);
+	        var context = new ConsultationsContext(_options, userService);
+	        var submitService = new SubmitService(context, _fakeUserService, _consultationService);
+			var commentService = new CommentService(context, _fakeUserService, authenticateService, submitService);
 
             // Act    
             var viewModel = commentService.GetCommentsAndQuestions(URI);
@@ -251,7 +265,9 @@ namespace Comments.Test.UnitTests
             AddCommentsAndQuestionsAndAnswers(ConsultationTwoURI, commentText, questionText, answerText, createdByUserId, (int)StatusName.Draft, _context);
             AddCommentsAndQuestionsAndAnswers(ChaperOverviewURI, "Another Users Comment", questionText, answerText, anotherUserId, (int)StatusName.Draft, _context);
 
-            var commentService = new CommentService(new ConsultationsContext(_options, userService), _fakeUserService, authenticateService, _consultationService);
+	        var context = new ConsultationsContext(_options, userService);
+	        var submitService = new SubmitService(context, _fakeUserService, _consultationService);
+			var commentService = new CommentService(context, _fakeUserService, authenticateService, submitService);
 
             // Act    
             var viewModel = commentService.GetCommentsAndQuestions(ConsultationOneURI, true);
