+++ Comments.Test/UnitTests/SubmitService.Tests.cs
@@ -27,14 +27,15 @@ namespace Comments.Test.UnitTests
 			var userService = FakeUserService.Get(isAuthenticated: true, displayName: "Benjamin Button", userId: userId);
 			var authenticateService = new FakeAuthenticateService(authenticated: true);
 			var consultationContext = new ConsultationsContext(_options, userService);
-			var commentService = new CommentService(consultationContext, userService, authenticateService, _consultationService);
+			var submitService = new SubmitService(consultationContext, userService, _consultationService);
+			var commentService = new CommentService(consultationContext, userService, authenticateService, submitService);
 
 			var locationId = AddLocation(sourceURI, _context);
 			var commentId = AddComment(locationId, "Comment text", false, userId, (int)StatusName.Draft, _context);
 
 			//Act
 			var commentsAndQuestions = commentService.GetCommentsAndQuestions(sourceURI, true);
-			var result = _consultationService.SubmitCommentsAndAnswers(new CommentsAndAnswers(commentsAndQuestions.Comments, new List<ViewModels.Answer>()));
+			var result = submitService.SubmitCommentsAndAnswers(new CommentsAndAnswers(commentsAndQuestions.Comments, new List<ViewModels.Answer>()));
 
 			var comment = commentService.GetComment(commentId);
 
@@ -62,8 +63,8 @@ namespace Comments.Test.UnitTests
 			var userService = FakeUserService.Get(isAuthenticated: true, displayName: "Benjamin Button", userId: userId);
 			var authenticateService = new FakeAuthenticateService(authenticated: true);
 			var consultationContext = new ConsultationsContext(_options, userService);
-
-			var commentService = new CommentService(consultationContext, userService, authenticateService, _consultationService);
+			var submitService = new SubmitService(consultationContext, userService, _consultationService);
+			var commentService = new CommentService(consultationContext, userService, authenticateService, submitService);
 			var answerService = new AnswerService(consultationContext, userService);
 
 			var locationId = AddLocation(sourceURI, _context);
@@ -73,7 +74,7 @@ namespace Comments.Test.UnitTests
 
 			//Act
 			var commentsAndQuestions = commentService.GetCommentsAndQuestions(sourceURI, true);
-			var result = _consultationService.SubmitCommentsAndAnswers(new CommentsAndAnswers(new List<ViewModels.Comment>(), commentsAndQuestions.Questions.First().Answers));
+			var result = submitService.SubmitCommentsAndAnswers(new CommentsAndAnswers(new List<ViewModels.Comment>(), commentsAndQuestions.Questions.First().Answers));
 
 			var answer = answerService.GetAnswer(answerId);
 
@@ -130,10 +131,11 @@ namespace Comments.Test.UnitTests
 			var userId = Guid.NewGuid();
 			var userService = FakeUserService.Get(isAuthenticated: true, displayName: "Benjamin Button", userId: userId);
 			var consultationContext = new ConsultationsContext(_options, userService);
+			var submitService = new SubmitService(consultationContext, userService, _consultationService);
 			AddSubmittedCommentsAndAnswers("consultations://./consultation/1/document/1/chapter/introduction", "Comment Text", "Question Text", "Answer Text", userId, consultationContext);
 
 			//Act
-			var actualResult = _consultationService.HasSubmittedCommentsOrQuestions(consultationSourceURI, userId);
+			var actualResult = submitService.HasSubmittedCommentsOrQuestions(consultationSourceURI, userId);
 			
 			//Assert
 			actualResult.ShouldBe(expectedResult);
@@ -155,10 +157,11 @@ namespace Comments.Test.UnitTests
 			var userId = Guid.NewGuid();
 			var userService = FakeUserService.Get(isAuthenticated: true, displayName: "Benjamin Button", userId: userId);
 			var consultationContext = new ConsultationsContext(_options, userService);
+			var submitService = new SubmitService(consultationContext, userService, _consultationService);
 			AddSubmittedCommentsAndAnswers("consultations://./consultation/1/document/1", "Comment Text", "Question Text", "Answer Text", userId, consultationContext);
 
 			//Act
-			var actualResult = _consultationService.HasSubmittedCommentsOrQuestions(consultationSourceURI, userId);
+			var actualResult = submitService.HasSubmittedCommentsOrQuestions(consultationSourceURI, userId);
 
 			//Assert
 			actualResult.ShouldBe(expectedResult);
@@ -180,10 +183,11 @@ namespace Comments.Test.UnitTests
 			var userId = Guid.NewGuid();
 			var userService = FakeUserService.Get(isAuthenticated: true, displayName: "Benjamin Button", userId: userId);
 			var consultationContext = new ConsultationsContext(_options, userService);
+			var submitService = new SubmitService(consultationContext, userService, _consultationService);
 			AddSubmittedCommentsAndAnswers("consultations://./consultation/1", "Comment Text", "Question Text", "Answer Text", userId, consultationContext);
 
 			//Act
-			var actualResult = _consultationService.HasSubmittedCommentsOrQuestions(consultationSourceURI, userId);
+			var actualResult = submitService.HasSubmittedCommentsOrQuestions(consultationSourceURI, userId);
 
 			//Assert
 			actualResult.ShouldBe(expectedResult);
