+++ Comments/Services/CommentService.cs
@@ -22,15 +22,15 @@ namespace Comments.Services
         private readonly ConsultationsContext _context;
         private readonly IUserService _userService;
         private readonly IAuthenticateService _authenticateService;
-	    private readonly IConsultationService _consultationService;
+	    private readonly ISubmitService _submitService;
 	    private readonly User _currentUser;
 
-        public CommentService(ConsultationsContext context, IUserService userService, IAuthenticateService authenticateService, IConsultationService consultationService)
+        public CommentService(ConsultationsContext context, IUserService userService, IAuthenticateService authenticateService, ISubmitService submitService)
         {
             _context = context;
             _userService = userService;
             _authenticateService = authenticateService;
-	        _consultationService = consultationService;
+	        _submitService = submitService;
 	        _currentUser = _userService.GetCurrentUser();
         }
 
@@ -124,9 +124,9 @@ namespace Comments.Services
 
 			var locations = _context.GetAllCommentsAndQuestionsForDocument(sourceURIs, isReview).ToList();
 
-		    var data = _consultationService.ConvertLocationsToCommentsAndQuestionsViewModels(locations);
+		    var data = ModelConverters.ConvertLocationsToCommentsAndQuestionsViewModels(locations);
 
-		    var consultationState = _consultationService.GetConsultationState(consultationSourceURI, locations);
+		    var consultationState = _submitService.GetConsultationState(consultationSourceURI, locations);
 
 			return new CommentsAndQuestions(data.comments, data.questions, user.IsAuthorised, signInURL, consultationState);
 	    }
