+++ Comments/Services/SubmitService.cs
@@ -1,99 +1,120 @@
-//using System;
-//using Comments.Models;
-//using Comments.ViewModels;
-//using System.Collections.Generic;
-//using System.Linq;
-//using Comments.Common;
+using System;
+using Comments.Models;
+using Comments.ViewModels;
+using System.Collections.Generic;
+using System.Linq;
+using Comments.Common;
 
-//namespace Comments.Services
-//{
-//	public interface ISubmitService
-//	{
-//		(int rowsUpdated, Validate validate) SubmitCommentsAndAnswers(CommentsAndAnswers commentsAndAnswers);
-//		bool HasSubmittedCommentsOrQuestions(string consultationSourceURI, Guid userId);
-//	}
-//	public class SubmitService : ISubmitService
-//    {
-//	    private readonly ConsultationsContext _context;
-//	    private readonly IConsultationService _consultationService;
-//	    private readonly User _currentUser;
+namespace Comments.Services
+{
+	public interface ISubmitService
+	{
+		(int rowsUpdated, Validate validate) SubmitCommentsAndAnswers(CommentsAndAnswers commentsAndAnswers);
+		bool HasSubmittedCommentsOrQuestions(string consultationSourceURI, Guid userId);
+		ConsultationState GetConsultationState(string sourceURI, IEnumerable<Models.Location> locations = null);
+	}
+	public class SubmitService : ISubmitService
+	{
+		private readonly ConsultationsContext _context;
+		private readonly IConsultationService _consultationService;
+		private readonly IUserService _userService;
+		private readonly User _currentUser;
 
-//	    public SubmitService(ConsultationsContext context, IUserService userService, IConsultationService consultationService)
-//	    {
-//		    _context = context;
-//		    _consultationService = consultationService;
-//		    _currentUser = userService.GetCurrentUser();
-//	    }
+		public SubmitService(ConsultationsContext context, IUserService userService, IConsultationService consultationService)
+		{
+			_context = context;
+			_consultationService = consultationService;
+			_userService = userService;
+			_currentUser = userService.GetCurrentUser();
+		}
 
-//	    public (int rowsUpdated, Validate validate) SubmitCommentsAndAnswers(CommentsAndAnswers commentsAndAnswers)
-//	    {
-//		    if (!_currentUser.IsAuthorised || !_currentUser.UserId.HasValue)
-//			    return (rowsUpdated: 0, validate: new Validate(valid: false, unauthorised: true, message: $"Not logged in submitting comments and answers"));
+		public (int rowsUpdated, Validate validate) SubmitCommentsAndAnswers(CommentsAndAnswers commentsAndAnswers)
+		{
+			if (!_currentUser.IsAuthorised || !_currentUser.UserId.HasValue)
+				return (rowsUpdated: 0, validate: new Validate(valid: false, unauthorised: true, message: $"Not logged in submitting comments and answers"));
 
-//			//if a user is submitting a different users comment, the context will throw an exception.
+			//if a user is submitting a different users comment, the context will throw an exception.
 
-//			var anySourceURI = commentsAndAnswers.SourceURIs.FirstOrDefault();
-//			if (anySourceURI == null)
-//				return (rowsUpdated: 0, validate: new Validate(valid: false, unauthorised: false, message: "Could not find SourceURI"));
+			var anySourceURI = commentsAndAnswers.SourceURIs.FirstOrDefault();
+			if (anySourceURI == null)
+				return (rowsUpdated: 0, validate: new Validate(valid: false, unauthorised: false, message: "Could not find SourceURI"));
 
-//		    var consultationState = _consultationService.GetConsultationState(anySourceURI);
+			var consultationState = GetConsultationState(anySourceURI);
 
-//			if (!consultationState.ConsultationIsOpen)
-//				return (rowsUpdated: 0, validate: new Validate(valid: false, unauthorised: false, message: "Consultation is not open for submissions"));
+			if (!consultationState.ConsultationIsOpen)
+				return (rowsUpdated: 0, validate: new Validate(valid: false, unauthorised: false, message: "Consultation is not open for submissions"));
 
-//			var hasSubmitted = HasSubmittedCommentsOrQuestions(anySourceURI, _currentUser.UserId.Value);
-//			if (hasSubmitted)
-//				return (rowsUpdated: 0, validate: new Validate(valid: false, unauthorised: false, message: "User has already submitted."));
+			var hasSubmitted = HasSubmittedCommentsOrQuestions(anySourceURI, _currentUser.UserId.Value);
+			if (hasSubmitted)
+				return (rowsUpdated: 0, validate: new Validate(valid: false, unauthorised: false, message: "User has already submitted."));
 
-//			var submissionToSave = _context.InsertSubmission(_currentUser.UserId.Value);
+			var submissionToSave = _context.InsertSubmission(_currentUser.UserId.Value);
 
-//		    var submittedStatus = _context.GetStatus(StatusName.Submitted);
-			
-//			UpdateCommentsModel(commentsAndAnswers.Comments, submissionToSave, submittedStatus);
-//		    UpdateAnswersModel(commentsAndAnswers.Answers, submissionToSave, submittedStatus);
+			var submittedStatus = _context.GetStatus(StatusName.Submitted);
 
-//			return (rowsUpdated: _context.SaveChanges(), validate: null);
-//		}
+			UpdateCommentsModel(commentsAndAnswers.Comments, submissionToSave, submittedStatus);
+			UpdateAnswersModel(commentsAndAnswers.Answers, submissionToSave, submittedStatus);
 
-//	    private void UpdateCommentsModel(IList<ViewModels.Comment> comments, Submission submission, Models.Status status)
-//	    {
-//			var commentIds = comments.Select(c => c.CommentId).ToList();
-//			_context.UpdateCommentStatus(commentIds, status);
+			return (rowsUpdated: _context.SaveChanges(), validate: null);
+		}
 
-//		    foreach (var commentInViewModel in comments)
-//		    {
-//			    commentInViewModel.UpdateStatusFromDBModel(status);
-//		    }
+		private void UpdateCommentsModel(IList<ViewModels.Comment> comments, Submission submission, Models.Status status)
+		{
+			var commentIds = comments.Select(c => c.CommentId).ToList();
+			_context.UpdateCommentStatus(commentIds, status);
 
-//			_context.AddSubmissionComments(commentIds, submission.SubmissionId);
-//		}
+			foreach (var commentInViewModel in comments)
+			{
+				commentInViewModel.UpdateStatusFromDBModel(status);
+			}
 
-//	    private void UpdateAnswersModel(IList<ViewModels.Answer> answers, Submission submission, Models.Status status)
-//	    {
-//			var answerIds = answers.Select(a => a.AnswerId).ToList();
-//			_context.UpdateAnswerStatus(answerIds, status);
+			_context.AddSubmissionComments(commentIds, submission.SubmissionId);
+		}
 
-//		    foreach (var answerInViewModel in answers)
-//		    {
-//			    answerInViewModel.UpdateStatusFromDBModel(status);
-//		    }
+		private void UpdateAnswersModel(IList<ViewModels.Answer> answers, Submission submission, Models.Status status)
+		{
+			var answerIds = answers.Select(a => a.AnswerId).ToList();
+			_context.UpdateAnswerStatus(answerIds, status);
 
-//		    _context.AddSubmissionAnswers(answerIds, submission.SubmissionId);
-//		}
+			foreach (var answerInViewModel in answers)
+			{
+				answerInViewModel.UpdateStatusFromDBModel(status);
+			}
 
-		
+			_context.AddSubmissionAnswers(answerIds, submission.SubmissionId);
+		}
 
+		public bool HasSubmittedCommentsOrQuestions(string anySourceURI, Guid userId)
+		{
+			if (string.IsNullOrWhiteSpace(anySourceURI))
+				return false;
 
-//	    public bool HasSubmittedCommentsOrQuestions(string anySourceURI, Guid userId)
-//	    {
-//		    if (string.IsNullOrWhiteSpace(anySourceURI))
-//			    return false;
+			var consultationsUriElements = ConsultationsUri.ParseConsultationsUri(anySourceURI);
 
-//			var consultationsUriElements = ConsultationsUri.ParseConsultationsUri(anySourceURI);
+			var consultationSourceURI = ConsultationsUri.CreateConsultationURI(consultationsUriElements.ConsultationId);
 
-//			var consultationSourceURI = ConsultationsUri.CreateConsultationURI(consultationsUriElements.ConsultationId);
+			return _context.HasSubmitted(consultationSourceURI, userId);
+		}
 
-//			return _context.HasSubmitted(consultationSourceURI, userId);
-//	    }
-//    }
-//}
+		public ConsultationState GetConsultationState(string sourceURI, IEnumerable<Models.Location> locations = null)
+		{
+			var currentUser = _userService.GetCurrentUser();
+			var consultationsUriElements = ConsultationsUri.ParseConsultationsUri(sourceURI);
+			var consultationDetail = _consultationService.GetConsultationDetail(consultationsUriElements.ConsultationId);
+
+			if (locations == null)
+			{
+				locations = _context.GetAllCommentsAndQuestionsForDocument(new[] { sourceURI }, isReview: true);
+			}
+
+			var hasSubmitted = currentUser != null && currentUser.IsAuthorised && currentUser.UserId.HasValue ? HasSubmittedCommentsOrQuestions(sourceURI, currentUser.UserId.Value) : false;
+
+			var data = ModelConverters.ConvertLocationsToCommentsAndQuestionsViewModels(locations);
+
+			var consultationState = new ConsultationState(consultationDetail.StartDate, consultationDetail.EndDate,
+				data.questions.Any(), data.questions.Any(q => q.Answers.Any()), data.comments.Any(), hasSubmitted);
+
+			return consultationState;
+		}
+	}
+}
